refactor: controller 입력받는 메서드 indent 줄이도록 수정
