+++ src/main/java/blackjack/domain/Card.java
@@ -3,8 +3,8 @@ package blackjack.domain;
 import java.util.Objects;
 
 public class Card {
-    private Suit suit;
-    private Rank rank;
+    private final Suit suit;
+    private final Rank rank;
 
     public Card(Suit suit, Rank rank) {
         this.suit = suit;
