+++ src/main/java/blackjack/domain/ScoreCalculator.java
@@ -1,5 +1,7 @@
 package blackjack.domain;
 
+import static blackjack.domain.WinnerDecider.MAX_SCORE;
+
 import java.util.List;
 
 public class ScoreCalculator {
@@ -19,7 +21,7 @@ public class ScoreCalculator {
         int min = Integer.MAX_VALUE;
         for (int score : scores) {
             int sum = maxDfs(depth + 1, totalScore + score, cards);
-            if (sum > 21) {
+            if (sum > MAX_SCORE) {
                 min = Math.min(min, sum);
                 continue;
             }
