+++ src/test/java/blackjack/domain/VictoryTest.java
@@ -1,5 +1,8 @@
 package blackjack.domain;
 
+import blackjack.domain.participants.Dealer;
+import blackjack.domain.participants.Player;
+import blackjack.domain.participants.Players;
 import java.util.List;
 import java.util.Map;
 import java.util.Stack;
